components moved to component directory
