+++ frontend/src/main/web/src/app/app.module.ts
@@ -4,12 +4,12 @@ import { HttpClientModule } from '@angular/common/http';
 import { AppRoutingModule, RoutingComponent } from './app-routing.module';
 
 import { AppComponent } from './app.component';
-import { EmployeeComponent } from './employee/employee.component';
-import { DepartmentComponent } from './department/department.component';
-import { EmployeeListComponent } from './employee-list/employee-list.component';
-import { DepartmentDetailsComponent } from './department-details/department-details.component';
-import { SalariesComponent } from './salaries/salaries.component';
-import { PageNotFoundComponent } from './page-not-found/page-not-found.component';
+import { EmployeeComponent } from './component/employee/employee.component';
+import { DepartmentComponent } from './component/department/department.component';
+import { EmployeeListComponent } from './component/employee-list/employee-list.component';
+import { DepartmentDetailsComponent } from './component/department-details/department-details.component';
+import { SalariesComponent } from './component/salaries/salaries.component';
+import { PageNotFoundComponent } from './component/page-not-found/page-not-found.component';
 
 import { EmployeeService } from './service/employee.service';
 import { DepartmentService } from './service/department.service';
@@ -35,4 +35,6 @@ import { SalariesService } from './service/salaries.service';
     bootstrap: [AppComponent]
 })
 
-export class AppModule { }
+export class AppModule { 
+    
+}
