+++ frontend/src/main/web/src/app/component/department-details/department-details.component.ts
