+++ frontend/src/main/web/src/app/component/employee-list/employee-list.component.html
@@ -1,0 +1,41 @@
+<!--The content below is only a placeholder and can be replaced.-->
+<div class="generic-container">
+    <div class="panel panel-default">
+        <div class="panel panel-default">
+            <!-- Default panel contents -->
+            <div class="panel-heading">
+                <span class="lead">List of Employees </span>
+            </div>
+            <div class="tablecontainer">
+                <table class="table table-hover">
+                    <thead>
+                        <tr>
+                            <th>Emp. Id</th>
+                            <th>First Name</th>
+                            <th>Last Name</th>
+                            <th>Birth Date</th>
+                            <th width="08%">Edit</th>
+                            <th width="08%">Remove</th>
+                        </tr>
+                    </thead>
+                    <tbody>
+                        <tr *ngFor="let employee of employees">
+                            <td><span>{{ employee.empNo }}</span></td>
+                            <td><span>{{ employee.firstName }}</span></td>
+                            <td><span>{{ employee.lastName }}</span></td>
+                            <td><span>{{ employee.birthDate }}</span></td>
+                            <td>
+                                <button type="button" class="btn btn-success custom-width"
+                                    (click)="editEmployee($event)"> Edit </button>
+                            </td>
+                            <td>
+                                <button type="button" class="btn btn-danger custom-width"
+                                    (click)="removeEmployee($event)">Remove</button>
+                            </td>
+                        </tr>
+                    </tbody>
+                </table>
+            </div>
+        </div>
+    </div>
+</div>
