+++ frontend/src/main/web/src/app/component/employee/employee.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { Employee } from '../model/employee';
-import { EmployeeService } from '../service/employee.service';
+import { Employee } from '../../model/employee';
+import { EmployeeService } from '../../service/employee.service';
 
 @Component({
     selector: 'app-employee',
