+++ frontend/src/main/web/src/app/component/salaries/salaries.component.html
