--- frontend/src/main/web/src/app/department-list/department-list.component.html
@@ -1,35 +1,0 @@
-<!--The content below is only a placeholder and can be replaced.-->
-<div class="generic-container">
-    <div class="panel panel-default">
-        <div class="panel panel-default">
-            <!-- Default panel contents -->
-            <div class="panel-heading">
-                <span class="lead">List of Department </span>
-            </div>
-            <div class="tablecontainer">
-                <table class="table table-hover">
-                    <thead>
-                        <tr>
-                            <th>ID.</th>
-                            <th>Department Name</th>
-                            <th width="08%">Edit</th>
-                            <th width="08%">Remove</th>
-                         </tr>
-                    </thead>
-                    <tbody>
-                        <tr *ngFor="let department of departments">
-                            <td><span>{{ department.deptNo }}</span></td>
-                            <td><span>{{ department.deptName }}</span></td>
-                            <td>
-                                <button type="button" class="btn btn-success custom-width"> Edit </button>
-                            </td>
-                            <td>
-                                <button type="button" class="btn btn-danger custom-width">Remove</button>
-                            </td>
-                        </tr>
-                    </tbody>
-                </table>
-            </div>
-        </div>
-    </div>
-</div>
