--- frontend/src/main/web/src/app/department-list/department-list.component.ts
@@ -1,22 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-import { Departments } from '../model/departments';
-import { DepartmentService } from '../service/department.service';
-
-@Component({
-  selector: 'app-department-list',
-    templateUrl: './department-list.component.html',
-    styleUrls: ['./department-list.component.css']
-})
-
-export class DepartmentListComponent implements OnInit {
-
-  departments: Departments[];
-  
-  constructor(private departmentService: DepartmentService){}
-
-  ngOnInit() {
-      this.departmentService.findAll().subscribe(data=> {
-          this.departments = data;
-      });
-  }
-}
